fix: bad dependency flow in staking
fix: move stakingTypes into wallet types

after: rebase fix

fix: more bad typing in staking

fix: more types fixies

chore: add script for File Watcher to fix Webstorm libs imports
