+++ packages/suite/src/actions/wallet/stake/stakeFormEthereumActions.ts
@@ -20,10 +20,11 @@ import {
     MIN_ETH_FOR_WITHDRAWALS,
     UNSTAKE_INTERCHANGES,
 } from '@suite-common/wallet-constants';
-import { ComposeActionContext, selectSelectedDevice } from '@suite-common/wallet-core';
+import { selectSelectedDevice } from '@suite-common/wallet-core';
 import { ethereumGetCurrentNonceThunk } from '@suite-common/wallet-core/src/send/sendFormEthereumThunks';
 import {
     AddressDisplayOptions,
+    ComposeActionContext,
     ExternalOutput,
     PrecomposedTransaction,
     PrecomposedTransactionFinal,
