+++ packages/suite/src/actions/wallet/stakeActions.ts
@@ -2,7 +2,6 @@ import { selectIsMevProtectionFeatureEnabled } from '@suite-common/mev';
 import { EarnFlow } from '@suite-common/suite-types/src/staking';
 import { notificationsActions } from '@suite-common/toast-notifications';
 import {
-    ComposeActionContext,
     addFakePendingCardanoTxThunk,
     replaceTransactionThunk,
     selectIsMevProtectionEnabled,
@@ -11,6 +10,7 @@ import {
     syncAccountsWithBlockchainThunk,
 } from '@suite-common/wallet-core';
 import {
+    ComposeActionContext,
     PrecomposedTransactionFinal,
     StakeFormState,
     StakeType,
