+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/UnstakeModal/UnstakeForm/UnstakeInputs.tsx
@@ -2,7 +2,6 @@ import { Translation, useTranslation } from '@suite/intl';
 import { useFormatters } from '@suite-common/formatters';
 import { formInputsMaxLength } from '@suite-common/validators';
 import { getNetworkDisplaySymbol } from '@suite-common/wallet-config';
-import { UnstakeFormState } from '@suite-common/wallet-core';
 import { getInputState, getStakingDataForNetwork } from '@suite-common/wallet-utils';
 import { Column, FractionButtonProps, Text } from '@trezor/components';
 import { InputWithOptions } from '@trezor/product-components';
@@ -20,6 +19,8 @@ import {
     validateFiatLimits,
     validateMin,
 } from 'src/utils/suite/validation';
+
+import { UnstakeFormState } from '../../../../../../wallet/stakeForm/StakeContext';
 
 export const UnstakeInputs = () => {
     const { translationString } = useTranslation();
