+++ packages/suite/src/types/wallet/claimForm.ts
@@ -1,7 +1,6 @@
 import { UseFormReturn } from 'react-hook-form';
 
-import { BaseStakeContextValues } from '@suite-common/wallet-core';
-import { StakeFormState } from '@suite-common/wallet-types';
+import { BaseStakeContextValues, StakeFormState } from '@suite-common/wallet-types';
 
 export type ClaimFormState = Omit<
     StakeFormState,
