+++ suite-common/staking/tsconfig.json
@@ -4,7 +4,6 @@
     "references": [
         { "path": "../wallet-config" },
         { "path": "../wallet-constants" },
-        { "path": "../wallet-core" },
         { "path": "../wallet-types" },
         { "path": "../wallet-utils" },
         { "path": "../../packages/connect" },
