+++ suite-common/wallet-core/src/index.ts
@@ -61,7 +61,6 @@ export * from './stake/stakeMiddleware';
 export * from './stake/stakeReducer';
 export * from './stake/stakeSelectors';
 export * from './stake/stakeThunks';
-export * from './stake/stakeTypes';
 export * from './token/stellarTokenThunks';
 export * from './transactions/transactionsActions';
 export * from './transactions/transactionsReducer';
