+++ suite-common/wallet-types/src/index.ts
@@ -9,6 +9,7 @@ export * from './sendForm';
 export * from './settings';
 export * from './selectedAccount';
 export * from './transaction';
+export * from './transactionReviewOutput';
 export * from './ethereumStaking';
 export * from './stakeForm';
 export * from './send';
@@ -18,3 +19,4 @@ export * from './globalSendReceive';
 export * from './device';
 export * from './baseCurrency';
 export * from './transactionSimulation';
+export * from './stakeTypes';
