+++ suite-common/wallet-types/src/transaction.ts
@@ -20,7 +20,6 @@ import {
 import { RequiredKey } from '@trezor/type-utils';
 
 import { Account, AccountDescriptor } from './account';
-import { FormStateTradingCryptoCurrency, FormStateTradingFiatCurrency } from './sendForm';
 
 export type { PrecomposedTransactionFinalCardano } from '@trezor/connect';
 
@@ -297,64 +296,6 @@ export type TransactionFiatRateUpdatePayload = {
 export type TransactionType = Pick<WalletAccountTransaction, 'type'>['type'];
 
 export type ExportFileType = 'csv' | 'pdf' | 'json';
-
-export type ReviewOutput =
-    | {
-          type:
-              | 'opreturn'
-              | 'data'
-              | 'locktime'
-              | 'fee'
-              | 'destination-tag'
-              | 'signing-with'
-              | 'network'
-              | 'timebounds'
-              | 'txid'
-              | 'address'
-              | 'amount'
-              | 'gas'
-              | 'contract'
-              | 'regular_legacy'
-              | 'approve_data'
-              | 'recipient_name';
-          label?: string;
-          value: string;
-          value2?: string;
-          token?: TokenInfo;
-          send?: undefined;
-          receive?: undefined;
-      }
-    | {
-          type: 'fee-replace';
-          label?: undefined;
-          value: string;
-          value2: string;
-          token?: undefined;
-          send?: undefined;
-          receive?: undefined;
-      }
-    | {
-          type: 'reduce-output';
-          label: string;
-          value: string;
-          value2: string;
-          token?: undefined;
-          send?: undefined;
-          receive?: undefined;
-      }
-    | {
-          type: 'traded_assets';
-          value: string;
-          value2: string;
-          label?: undefined;
-          token?: undefined;
-          send: FormStateTradingCryptoCurrency;
-          receive: FormStateTradingCryptoCurrency | FormStateTradingFiatCurrency;
-      };
-
-export type ReviewOutputType = ReviewOutput['type'];
-
-export type ReviewOutputState = 'active' | 'success' | undefined;
 
 export type ExcludedUtxos = Record<string, 'low-anonymity' | 'dust' | undefined>;
